[frontendGraph] Simplify SimpleGraphCellPainter and rename to GraphCellPainter, increase BranchOrCommitCellRenderer performance (closes #98)
